# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hotel & Bar Management API
  version: 0.1.0
  description: Tenant-scoped hospitality API. Authenticated calls use **Bearer** JWT.
    Send header **X-Tenant-Id** (tenant UUID) on tenant-scoped routes; omit on `/health/`,
    `/auth/token/`, `/auth/token/refresh/`, `/auth/password/reset/`, `/auth/password/reset/confirm/`,
    and `/invites/accept/`.
paths:
...
  /api/v1/auth/token/:
    post:
      operationId: auth_token_create
      description: |-
        Takes a set of user credentials and returns an access and refresh JSON web
        token pair to prove the authentication of those credentials.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenObtainPairRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenObtainPair'
          description: ''
...
  /api/v1/integrations/payment-webhook/:
    post:
      operationId: integrations_payment_webhook_create
      description: |-
        ``POST`` with JSON body containing at least **event_id**, **provider**, **status**.
        Requires header **X-Tenant-Id** and **X-Webhook-Secret** matching
        ``PAYMENT_WEBHOOK_SHARED_SECRET`` when that setting is non-empty.
        Idempotent per tenant + **event_id** (~30 days).
      parameters:
      - in: header
        name: X-Tenant-Id
        schema:
          type: string
          format: uuid
        required: true
      - in: header
        name: X-Webhook-Secret
        schema:
          type: string
        required: true
      tags:
      - integrations
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '403':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/orders/{id}/payments/:
    post:
      operationId: orders_payments_create
      parameters:
      - name: X-Tenant-Id
        in: header
        required: false
        description: Active tenant UUID. Required for most authenticated, tenant-scoped
          routes.
        schema:
          type: string
          format: uuid
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this order.
        required: true
      tags:
      - orders
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderRead'
          description: ''
...
components:
  schemas:
...
    KdsStatusEnum:
      enum:
      - pending
      - in_prep
      - ready
      - served
      type: string
      description: |-
        * `pending` - Pending
        * `in_prep` - In prep
        * `ready` - Ready
        * `served` - Served
...
    MethodEnum:
      enum:
      - cash
      - card
      - other
      type: string
      description: |-
        * `cash` - Cash
        * `card` - Card
        * `other` - Other
...
    OrderLineRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        menu_item:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        label:
          type: string
          readOnly: true
        quantity:
          type: string
          format: decimal
          pattern: ^-?\d{0,7}(?:\.\d{0,3})?$
          readOnly: true
        unit_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          readOnly: true
        line_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
          description: Line subtotal before tax (qty × unit price).
        tax_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        sort_order:
          type: integer
          readOnly: true
        is_voided:
          type: boolean
          readOnly: true
        void_reason:
          type: string
          readOnly: true
        voided_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        kds_station:
          type: string
          readOnly: true
        kds_status:
          allOf:
          - $ref: '#/components/schemas/KdsStatusEnum'
          readOnly: true
        modifiers_snapshot:
          readOnly: true
          description: Chosen modifier options at sale time (name + price_delta per
            option).
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - is_voided
      - kds_station
      - kds_status
      - label
      - line_total
      - menu_item
      - modifiers_snapshot
      - quantity
      - sort_order
      - tax_amount
      - unit_price
      - void_reason
      - voided_at
...
    OrderRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        bill_reference:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/PosOrderStatus'
          readOnly: true
        table_id:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        table_label:
          type: string
          readOnly: true
        outlet_id:
          type: string
          format: uuid
          readOnly: true
        outlet_name:
          type: string
          readOnly: true
        currency:
          type: string
          readOnly: true
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        tax_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        discount_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
          description: Order-level discount subtracted after line subtotals and tax
            (open orders only).
        total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        amount_paid:
          type: string
          readOnly: true
        balance_due:
          type: string
          readOnly: true
        is_paid:
          type: boolean
          readOnly: true
        folio:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        applied_promotion:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineRead'
          readOnly: true
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
          readOnly: true
      required:
      - amount_paid
      - applied_promotion
      - balance_due
      - bill_reference
      - created_at
      - currency
      - discount_amount
      - folio
      - id
      - is_paid
      - lines
      - outlet_id
      - outlet_name
      - payments
      - status
      - subtotal
      - table_id
      - table_label
      - tax_total
      - total
      - updated_at
...
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        order_id:
          type: string
          format: uuid
          readOnly: true
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        method:
          allOf:
          - $ref: '#/components/schemas/MethodEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - amount
      - created_at
      - id
      - method
      - order_id
    PosOrderStatus:
      enum:
      - open
      - cancelled
      - closed
      type: string
      description: |-
        * `open` - Open
        * `cancelled` - Cancelled
        * `closed` - Closed
...
    TokenObtainPair:
      type: object
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
          readOnly: true
      required:
      - access
      - refresh
    TokenObtainPairRequest:
      type: object
      properties:
        email:
          type: string
          writeOnly: true
          minLength: 1
        password:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - email
      - password
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantId:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: Active tenant UUID; required for most authenticated routes.
servers:
- url: http://127.0.0.1:8000
  description: Local
- url: https://api.example.com
  description: Production (example)
...